Update Assignment EX6.c

+++ C Language/C Basics Assignment/Assignment EX6.c
@@ -8,19 +8,19 @@
 #include <stdlib.h>
 
 int main() {
-	int a,b,temp;
+	float a,b,temp;
 	printf("Enter value of a: ");
 	fflush(stdin); fflush(stdout);
-	scanf("%d",&a);
+	scanf("%f",&a);
 	printf("Enter value of b: ");
 	fflush(stdin); fflush(stdout);
-	scanf("%d",&b);
+	scanf("%f",&b);
 	fflush(stdin); fflush(stdout);
 	temp =a;
 	a=b;
 	b= temp;
-	printf("After swapping, value of a = %d \n",a);
-	printf("After swapping, value of b = %d \n",b);
+	printf("After swapping, value of a = %f \n",a);
+	printf("After swapping, value of b = %f \n",b);
 	fflush(stdin); fflush(stdout);
 	return 0;
 }
